lamnt98 commit move function 28/06
