--- project/src/main/java/com/example/webDemo3/controller/DemoController.java
@@ -1,4 +1,0 @@
-package com.example.webDemo3.controller;
-
-public class DemoController {
-}
