+++ project/src/main/java/com/example/webDemo3/controller/UserController.java
@@ -2,9 +2,9 @@ package com.example.webDemo3.controller;
 
 import com.example.webDemo3.dto.LoginResponseDto;
 import com.example.webDemo3.dto.MessageDTO;
-import com.example.webDemo3.dto.request.LoginRequestDto;
-import com.example.webDemo3.entity.User;
-import com.example.webDemo3.service.LoginService;
+import com.example.webDemo3.dto.ViewPerInforResponseDto;
+import com.example.webDemo3.dto.request.*;
+import com.example.webDemo3.service.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -15,10 +15,46 @@ import javax.servlet.http.HttpSession;
 
 @RestController
 @RequestMapping("/api/user")
-public class LoginApiController {
+public class UserController {
+
+    @Autowired
+    private ChangePasswordService changePasswordService;
+
+    @Autowired
+    private ViewPerInfoService viewPerInfoService;
+
+    @Autowired
+    private EditPerInforService editPerInforService;
 
     @Autowired
     private LoginService loginService;
+
+    /**
+     * lamnt98
+     * 23/06
+     * catch request from client to change password
+     * @param model
+     * @return MessageDTO
+     */
+    @PostMapping("/changepassword")
+    public MessageDTO login(@RequestBody ChangePasswordRequestDto model)
+    {
+        return changePasswordService.checkChangePasswordUser(model);
+    }
+
+
+    /**
+     * lamnt98
+     * 26/06
+     * catch request from client to edit information of user
+     * @param model
+     * @return MessageDTO
+     */
+    @PostMapping("/editinformation")
+    public MessageDTO login(@RequestBody EditPerInforRequestDto model)
+    {
+        return editPerInforService.editUserInformation(model);
+    }
 
     /**
      * kimpt142
@@ -28,7 +64,7 @@ public class LoginApiController {
      * @return LoginDto with (1,success) if success
      */
     @PostMapping("/login")
-    public LoginResponseDto login(@RequestBody LoginRequestDto model,  HttpSession session)
+    public LoginResponseDto login(@RequestBody LoginRequestDto model, HttpSession session)
     {
         LoginResponseDto responseDto = loginService.checkLoginUser(model);
         if(responseDto.getMessage().getMessageCode() == 0) {
@@ -56,4 +92,20 @@ public class LoginApiController {
         }
         return message;
     }
+
+
+    /**
+     * lamnt98
+     * 25/06
+     * catch request from client to find information of user
+     * @param model
+     * @return ViewPerInforResponseDto
+     */
+    @PostMapping("/viewinformation")
+    public ViewPerInforResponseDto viewInformation(@RequestBody ViewPerInforRequestDto model)
+    {
+        ViewPerInforResponseDto responseDto = viewPerInfoService.getUserInformation(model);
+        return responseDto;
+    }
+
 }
